refactor(resultset): Extract InMemoryMetadataResultSet from MultiSchemaMySqlMetadataHandler
* Move the private nested InMemoryMetadataResultSet proxy out to its own
  public class in org.dbunit.database so other IMetadataHandler
  implementations can reuse it, not just MySQL's.
* No behavior change; MultiSchemaMySqlMetadataHandler's merge() usage is
  unaffected.

Refs: 923
